requirements.txt
streamlit>=1.45.0
pandas
gspread>=6.0.0
google-auth>=2.0.0
plotly
streamlit-aggrid
